БЕЗОПАСНЫЙ ТРЕК: Исправление narrowing conversions
- Исправлены 5 из 5 narrowing conversions
- src/fake_sensor.cpp:27 - приведение random() к float
- src/modbus_sensor.cpp:513 - приведение uint16_t к float в convertRegisterToFloat()
- src/ota_manager.cpp:73,144,604 - замена int на size_t для сравнений с strlen()
- Улучшена безопасность типов и предотвращены потери данных
- Все тесты проходят успешно

Технический долг: 25 (Low) - отличное состояние

+++ src/modbus_sensor.cpp
@@ -510,7 +510,7 @@ void readSensorData()
 
 float convertRegisterToFloat(uint16_t value, float multiplier)
 {
-    return value * multiplier;
+    return static_cast<float>(value) * multiplier;
 }
 
 /**
